requirements.txt
# ══════════════════════════════════════════════════════════════════
#  Smart City ANPR System — Python Dependencies (Sprint 1-3)
#  SRM Institute of Science and Technology, 2024–25
#  Install:  pip install -r requirements.txt
# ══════════════════════════════════════════════════════════════════

# Core computer vision
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0

# Vehicle detection (YOLOv8/v9)
ultralytics>=8.1.0

# OCR
easyocr>=1.7.0

# GenAI super-resolution (Real-ESRGAN)
basicsr>=1.4.2
facexlib>=0.3.0
gfpgan>=1.3.8
realesrgan>=0.3.0

# Deep learning runtime
torch>=2.0.0
torchvision>=0.15.0

# Face anonymisation
mediapipe>=0.10.0

# REST API (Sprint 3)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
httpx>=0.26.0

# Alerts — optional (SMS via Twilio)
twilio>=8.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0

# Utilities
requests>=2.31.0
python-dotenv>=1.0.0
tqdm>=4.66.0
